docker-build: fix building for macos-amd64 by forcing the use of an old brew
Fix building for macos-amd64 by forcing theuse of an old brew:

- use `brew-install` from an old reference
- set `HOMEBREW_NO_INSTALL_FROM_API=1` so the local repository is used with old references

Darling emulates an old macOS, and then fetches an old xcode.
Current brew doesn't support that old macOS, neither that old xcode.

We may want to build on older macOS the same way we do with older Debian
for Linux anyway.

We use brew to install p7zip and cmake, we don't need it for anything else
since we build the external_deps ourselves.

In the future we may rely on cmake binary tarball from cmake.org and build p7zip ourselves
but using an old brew (and make sure it doesn't auto-update) is good enough to fix the build.

We now use a subscript so environment variables are set within darling and not outside of it.