docker: make possible to only build the engine without committing it to Unvanquished
One can only build the engine without committing it to Unvanquished by doing:

./docker-build --engine-reference=<git_reference> --targets=<engine_target>

The reference for the game is now optional, the current master branch
will be built if not set. The currently committed engine reference for
the game reference will be built if the engine reference is not set.

The repository clone is now complete, instead of doing a shadow clone.
This makes easier to checkout arbitrary commits, and in the future
this will ensure the most recent tag is reachable, to compute the
version string.